Refactor the core matching code.
This can then be used by systems wanting to do the same thing but internally. For instance, our internal search scoring uses this to bypass the HTTP & Django stack overhead.

It would be nice to pull even more of this out, probably into a Strategy pattern; there's still a lot of duplicated code :-/